test: replace dd and truncate with fallocate
Replaced all instances of dd or truncate which
prepare file to be used in tests.
This should speed up the tests atleast a little bit
and make them not depend on speed of underlying device.

'dd' takes time to write data to device, taking the time.
'truncate' allocates a sparse file, which in some cases
may cause issues when we run out of the disk space.
'fallocate' does not write any data, and makes actual
reservation of space in the filesystem.

NOTE: This patch does not touch instances where dd is used
as part of the test for clearing data on device or preparing
certain files with random data.

Change-Id: Ic5a1360743fdbf461137637ba030524c9aacf680
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>