Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Ziye, Ben,
> 
> I thought Ben's suggestion may be like (i.e., replace 50 count to time based and cache tsc value to spdk thread).
> 
> spdk_nvmf_tcp_poll_group_poll(group)
> 	thread = spdk_get_thread();
> 	now = thread->now;
> 
> 	if (now < tgroup->restart_time)
> 		return 0;
> 
> 	rc = spdk_sock_group_poll(group);
> 	if (rc < 0)
> 		return rc;
> 	else if (rc == 0)
> 		tgroup->restart_time = now + spdk_max(0, 50us - (now - tgroup->busy_start));
> 	else {
> 		if (now > tgroup->busy_start)
> 			tgroup->busy_start = now;
> 	}
> 
> 	return 0
> 
> and storing tsc value to spdk_thread and referring it in poller is not so complex.

That's exactly what I was thinking. Things that may be too slow in this are calling spdk_get_ticks() (but we can cache that in the thread) and spdk_get_thread() (but we can cache the thread in the group).

Patch-set: 1
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>