nvme: Add new storage pool allocator to support multiple users of CMB
This enhancement implements a simplistic memory pool storage allocator
to support multiple uses of a device's NVMe Controller Memory
Buffer (CMB).  This keeps track of the CMB memory area, supporting
both dynamic allocation and deallocation.

This change is transparent to SPDK applications, that is, there are
no changes to the SPDK user interface.  This functionality is for
PCIe attached NVMe devices only.

This addresses Trello entry
https://trello.com/c/dYaEhJCX/3-add-cmb-allocator

Signed-off-by: James Bergsten <jamesx.bergsten@intel.com>
Change-Id: I477559a551d85f7e662ce00907b654daeb4527e5