test/lvol: rewrite destroy_lvol_store_positive to bash
Those test cases were basically testing hotremove, so
that's how we're going to name them now.

Note that in the first test case we're getting the list
of all lvolstores and lvols and checking if it's empty,
but then we're not doing it in other cases - that's because
the run_test wrapper function generally checks that for
us automatically, although in the first case we wanted
to check it before we removed the malloc.

Change-Id: Ib79ab7e50e3022df981f7092436b0a0991d5c9b9
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>