nvmf/tcp: cleanup in spdk_nvmf_tcp_qpair_flush_pdus_internal
Purpose: We do not need call this function, it causes the
bug during shutdown test. Since it will call the function
recursively, and it has no side effects to remove calling
this function.

Test case:  Using the following command to test
./test/nvmf/target/shutdown.sh --iso --transport=tcp
without this patch, it will cause coredump.

Also, re-enable the shutdown.sh tests on autotest so that we can confirm
this change will not break the pool.

Change-Id: Ifa193faa3f685429dcba7557df5b311bd566e297
Signed-off-by: Ziye Yang <ziye.yang@intel.com>