Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Thanks Ziye, This definitely fixes the coredump I was observing. I ran the shutdown test about 20 times on this patch. I still don't exactly understand why this fixes it though. Is the flush_pdus_internal function able to be called from a different thread than the sock group and that's what was causing the corruption of the sock group?

I am good to merge this if you have confidence that this is the actual cause of the corruption of the TAILQ and not just changing the timing so that we don't hit the assert.
I just have one small nit (below) to change before it gets merged.

Patch-set: 1
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1