Update patch set 5
Patch Set 5: Code-Review-1

We tested this in the following scenario:
1. Create 2 subsystems, A, B
2. Add hostA to subsystem A
3. Add hostB to subsystem B
4. Discovery from hostA - discovery page contains only subsystem A
5. Discovery from hostB - discovery page contains subsystem A, where we expected to see subsystem B

From another review of the code, we found that this is because the generation number does not change between discoveries (since we create the subsystems prior to the discover commands).

A simple fix for this is to save the previous hostnqn, and add to the condition for calling nvmf_update_discovery_log: hostnqn != previous hostnqn, e.g.

        if (tgt->discovery_log_page == NULL ||
            tgt->discovery_log_page->genctr != tgt->discovery_genctr ||
            tgt->discovery_log_page->prev_hostnqn != hostnqn) {
                nvmf_update_discovery_log(tgt, hostnqn);
        }

This is a simple fix, but may an alloc/free whenever a we get a request basically.

Patch-set: 5
Reviewer: Gerrit User 1014226 <1014226@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1