lib/notify: lockless notifications
This is lockless implementation of data ring with random access to be
used as notification bus.

PROBLEMS:

 1) Multiple threads are accessing the same ring.

 2) While puting notification to the ring takes some time, we cannot
    be sure that notifications data is ready for consumers in the same
    order they come.

    It can lead to fragmentation which makes difficult to fetch a
    continuous sequence of notifications and makes random access to
    the ring non-trivial.

 3) Copying notification data to the ring is not atomic, so
    multithreaded access can cause data corruption.

SOLUTION:

We can use two rings, one for data (input ring) and second (output ring)
to expose notifications for the consumer.

When producer sends a notification it is put to the data ring, and when
is ready to consume, its pointer is put to the output ring. It guarantee
continuous sequence of notifications in output ring and is much more
suitable for random access.

Data ring should be larger than output ring to protect notifications
being exposed.

Producer:
 1) Producer sends a notification.
 2) Entry for notification is allocated in the data queue.
 3) Data is copied to the allocated space.
 4) When notification is ready, put its pointer to the output ring. It
    will provide continuous sequence of notification with valid data.

Consumer:
 1) Consumer requests notification by index.
 2) Validate if index is in range of currently exposed notifications.
 3) Copy notification data to the local buffer.
 4) Revalidate index to check if data could be overwriten by another
    thread during 3). If so, reject notification.

Change-Id: Ic4f9eb2622e30e1b63a0f8a2e504e13b888d90e7
Signed-off-by: Tomasz Kulasek <tomaszx.kulasek@intel.com>