Update patch set 2
Patch Set 2:

All right then.
I don't think we need to add any special support to the script, right? It's just kind of a different FIO binary to which we can point using --fio-bin option in the scripts. Is that correct?

Vishal, John - Can I ask for some tips how to enable/run this fio-t on my server?

Patch-set: 2