githooks: limit the number of threads for pre-push hook
The script probably meant to execute make with $(nproc),
but executed it with ${nproc} instead. "nproc" was not
defined, evaluated to nothing, and -j without integer
spawned unlimited number of processes, which was slow.

While here, also use sysctl -a | grep -E -i 'hw.ncpu'
to get the number of cores on BSD. That's what we already
do in autobuild.sh.

Fixes #881

Change-Id: I8b07a2c28c4834b5dfb1c1bfa66d2b696d85720f
Reported-by: Jan Kryl <jan.kryl@mayadata.io>
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>