Update patch set 9
Patch Set 9:

OK.  Will need to get new versions of these patches in the next day or so if we still want to get them in for 19.07.  Or we could just decide now that we'll hold off until 19.10.

Patch-set: 9
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>