Update patch set 16
Patch Set 16: Code-Review+1

This change removes a couple of paths that I see as being able to recurse in undesirable ways.  The main one in my mind is calling back into process_pending through the pdu callback. Ziye first pointed out this recursion to me during our 1:1 yesterday.
The path I see as possible is this:
1. Call process_pending and enter the tailq_foreach_safe loop on the pending buffers list. We start with the first entry in the list and the TAILQ_FOR_EACH_SAFE macro stores a pointer to the second entry in the list as the next variable. 
2. Find a buffer remove the first entry from the list and call through the following stack:
- spdk_nvmf_tcp_pdu_set_buf_from_req
- spdk_nvmf_tcp-send_r2t_pdu
- spdk_nvmf_tcp_qpair_write_pdu
- spdk_nvmf_tcp_qpair_flush_pdus
- spdk_nvmf_tcp_qpair_flush_pdus_internal
- spdk_nvmf_tcp_qpair_process_pending
3. Here we again enter a tailq_foreach_safe loop on the pending buffers list. However this time, the first entry has already been removed. So we start at the second entry.
4. We process several requests, removing each of them from the list, and then return back up through the stack.
5. When we return back to the initial for loop from step 1, the original second entry is still stored in the tvar variable (even though it has already been removed from the tailq). When we try to process this entry, it is in some unknown state. Most of the time this is fine since there are several states that the qpair can be in that we don't do anything with. However, in the rare case, like when a qpair is in an error state during shutdown we end up catching an entry that has already been completed and freed. That is when we hit the assert in the shutdown case.
I am not 100% sure that calling TAILQ_FOREACH_SAFE recursively is not allowed, but from looking at the macros I believe that it is dangerous and could result in the corruption the Ziye is trying to avoid in this patch and that I detailed above.

Patch-set: 16
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1