nvme: only add 2s delay when iommu is enabled
This delay works around an issue with older Intel SSDs
that couldn't undergo initialization immediately after an
FLR (which VFIO does during initialization).  When the
iommu is disabled, we know we aren't using VFIO, so we can
skip the delay.

Future patches will add quirks so that this delay is only
incurred on the SSDs that require it.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I2854af556b67a9a4e5f686c72c4407962808f963