Revert "test/setup: check open files limit"
On multiple development systems (including mine), this results
in new error messages when running scripts/setup.sh.  This has
the potential of causing lots of questions when users upgrade
to SPDK 19.07.

I understand the desire for the error messages, but I think we have
to come up with a smarter check than just against "unlimited".

This reverts commit 0abee610f04b91a326fa9dbbb2d596b67221fd06.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I2854af556b67a9a4e5f686c72c4407962808f964