Update patch set 9
Patch Set 9: Code-Review-1

So this patch looks good - the only thing we're missing is some kind of test for this new functionality.  Not saying we need to verify the results that it returns, but we should at least execute this code somewhere in our nvmf tests to make sure the functionality doesn't result in crashes, etc.  We didn't have an nvmf_get_stats call prior to this patch series, but we really need one especially with all of the extra functionality added in this patch series.

So marking this patch as -1 just to make sure we don't merge these patches until there's a test in place.  I'm fine with adding a patch at the end of this series that adds an nvmf_get_stats call in one of the existing nvmf test scripts.

Patch-set: 9
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1