Update patch set 3
Patch Set 3: Code-Review-1

I looked over the usage and I don't think this is going to work without some more thought. There are a number of places where these values are referenced at run time. Specifically, action_on_timeout and timeout_us is used on the timeout path from multiple threads. The two poll periods are only used when the device is first created today, but if we ever wanted to dynamically start and stop pollers they'd be a problem. And finally the retry_count is the trickiest of all because that needs to get assigned to the global spdk_nvme_retry_count, but that value is getting used from within the main I/O path.

I think for this to work we need to make a copy of these values and place them into the bdev. When the user updates the "default" version of these opts via RPC, it should only apply to new bdevs created (by copying the opts into the bdev on creation).

The hardest part of that copying strategy is fixing spdk_nvme_retry_count, which needs to be a cached value in the spdk_nvme_ctrlr or spdk_nvme_qpair I think.

Patch-set: 3
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1