Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Code does not contain FC Transport.

Sorry if this is causing some confusion. Now that FC NVMe is committed to SPDK master we enabled SPDK CI to start running tests on Gerrit postings and post review result. As patch 1 does not contain FC NVMe code, it just posted review with a message stating FC transport is not present.

Patch-set: 2
CC: Gerrit User 1015300 <1015300@d5d70762-12d0-45a1-890d-524b12d3f735>