Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-2
> 
> This patch effectively breaks VFIO no-iommu on older drives.

If VFIO no-iommu mode is important, we should probably add it to our CI.

We could also change this to only do the delay if the controller is attached to VFIO.  That was actually the original route I was going to take, until I saw that env already exported the iommu function.

> 
> I believe it's still possible to do an FLR for UIO-bound device via sysfs - /sys/bus/pci/devices/<bdf>/reset

If people are doing resets via sysfs, I think we have more to worry about than just this 2s delay.

> 
> Can we go straight for the SSD quirks?

We can do SSD quirks, but there's a risk that it will break other SSDs if we only enable the quirk for specific older Intel models.  I'm still planning to do the quirks, but think that will be a longer-term solution so we can gather feedback from the community specifically on non-Intel drives.

Patch-set: 2
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>