Update patch set 2
Patch Set 2: Code-Review+1

I think this is OK but am wondering if we should include a link to the patch.  I'm just thinking that if I'm an SPDK consumer I'd want to know exactly what kind of request and how it was fixed in case I don't want to upgrade but instead apply the fix on my own.  I'm sure someone would ask if that were the situation but if you all agree I think including a link to the patch would be good.

Patch-set: 2
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1