scripts/pkgdep.sh: add shellcheck installation.
This will enable installation of Shellcheck using a precompiled
binary file, which should work on most Linux distributions.
This allows us to save time and space and not install cabal-install
dependency, which we would never use again.

Tested the binary in a VM with Fedora 29, Ubuntu 16 and Centos 7.

Reason for this way of installation instead of package manager is
latest (0.7.0) Shellcheck release, which is still not available on
majority of distibutions, but has some interesting new features
we can use.

Change-Id: I654f4168d5fab55dc7e1eeaa353ea6da24d04dd7
Signed-off-by: Karol Latecki <karol.latecki@intel.com>