test/compress: enable QAT testing in CI
Just for a quick test of QAT in the CI systems I've made
per patch use QAT if available and nightly use ISAL. It would
be good to do more with QAT nightly but due to the way the PMD
options are set it's a little more work to test both (and takes
more time as each PMD requires a new instance of the test app).

So, this is WIP to make sure QAT works then I'll experiment with
how long it takes to test w/both nightly.

Change-Id: If4bde8ed4b25363da668ab6084c67af4e95ec66d
Signed-off-by: paul luse <paul.e.luse@intel.com>