test/compress: enable QAT testing in CI
We need to test both ISAL and QAT but this would be very time
consuming in CI to test both per patch and nightly.  So, for
per patch we use QAT if its there and ISAL if not.  For nightly,
we run bevio and bdevperf twice, once with ISAL and once with
QAT to assure we have nightly coverage of both.  To save time,
the FS test is still only run nightly but just once using the
same auto-detect default feature as per patch.

Change-Id: If4bde8ed4b25363da668ab6084c67af4e95ec66d
Signed-off-by: paul luse <paul.e.luse@intel.com>