Update patch set 1
Patch Set 1:

Another comments:

Even if we use io_submit, for read/write, we still get part of data. For example, if we read 4096 bytes, it may be return 200 bytes or other values < 4096. So it is not exactly align with block I/O. And we will need to try the read again, so it will make the upper layer programming much more complicated.

Patch-set: 1
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>