nvme: add an option 'no_shn_notification' to driver
We will do a normal shutdown notification for most cases, and
it will take some time e.g. 2 seconds to take the action.  If
user environment has several drives, because the process is
synchronous, so the time maybe longer that user can accept.

Since user can know exactly what they would like to do next,
so here we provide an option to user, user can set it to true
to skip the shutdown notification process so that they can
have very quick startup.

Change-Id: Ie7f87115d57776729fab4cdac489cae6dc13511b
Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>