Update patch set 7
Patch Set 7:

Hi, just want to make sure that I am understanding this correctly. Before, we were operating under the assumprion that there was only one io_channel per underlying device so when we freed the raid bdev, it was retroactively freeing the remaining channels causing a race where we touched the raid bdev after it was free. And now, we are making sure that we put all of the I/O channels before destroying the raid bdev.

Is that correct?

Patch-set: 7
CC: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>