lib/reduce: remove req from executing list before callback
The allocated request is still safe as it not placed on the free
list until after it is used to determine if there is a queued
overlapped request for the logical map index associated with it.

The benefit in reversing this order is that a back to back
sequential operation will not be queued unnecessarily.

Bdevperf w/verify is an example of an app that produces this
workload.

Signed-off-by: paul luse <paul.e.luse@intel.com>
Change-Id: I98099d299c9ab906d115b905b7f2c86951045046