mk: move the blob_bdev lib under bdev.
While the blob_bdev doesn't fit nicely into the bdev/vbdev module
abstractions, it contains similar dependencies as the other bdev
modules. It also creates an unresolvable dependency chain when linking
shared libraries as is. When placed inside the blob directory, we would
be required to build the bdev directory before blob_bdev, but the lvol
bdev inside of the bdev directory depends on blob_bdev creating a
deadlock.

Change-Id: I47550df4e1c4c2d9abf7620e86a95ced67b4f46c
Signed-off-by: Seth Howell <seth.howell@intel.com>