nvmf/tcp: Change pdu hdr to hold pointers
Instead of the PDU holding the actual header, make
it a pointer instead.

In addition, have some raw memory still available in the
PDU that the pointers just point at. For now, all uses
of the PDU set up the header pointer to point at this
raw memory, so there isn't actually a functional change
from before.

The goal is to set up the PDU so that it can, in the future,
point at another location (like the big recv buffer).

Change-Id: I5330bd0ff8292e16fab2e74a7b431e2eafe1079e
Signed-off-by: Ben Walker <benjamin.walker@intel.com>