bdev/compress: enable force deletion of a comp bdev w/o pmem file
A new parameter was added to the RPC for deleting a compress bdev,
delete_compress_bdev(). The new parameter `force` will force deletion
of the compressed volume even if the pm metadata file cannot be found.
Previously, if the pm metadata file was deleted, there was no way to
delete the compressed volume.

Due to how the reduce vol deletion was designed, changes were made
To enable a new ‘mode’ of loading a compressed volume so that it can
be forcibly deleted. A comp vol that is loaded without a pmem file will
have 0 blocklen and only RPC’s will be allowed.

fixes issue #890

Change-Id: If1debc4751461c37f9b4008609ead08c0d2e5746
Signed-off-by: paul luse <paul.e.luse@intel.com>