split: Add support for multiple times config
Previously, split vbdev can only be configured once no
matter whether we use all the space of the base bdev
or not, which is not efficient enough.

Adding support for multiple times configuration. Base
bdev can be split multiple times until we have used all
the spaces. And the size of split vbdev can various for
the same base bdev.

Change-Id: I5c44a9b99f20c68f0d84f23ff5f1d5ba73b46b74
Signed-off-by: Chunyang Hui <chunyang.hui@intel.com>